Update Post feature
1: modified update function to accept new updates (views>Post>Update.php)
2: added updatefiled function to accept new photos in the post (models>post.php)

+++ models/post.php
@@ -52,27 +52,28 @@ class Post {
         $req->bindParam(':title', $title);
         $req->bindParam(':content', $content);
         $req->bindParam(':image', $image);
-
 // set name and price parameters and execute
+
         if (isset($_POST['title']) && $_POST['title'] != "") {
             $filteredTitle = filter_input(INPUT_POST, 'title', FILTER_SANITIZE_SPECIAL_CHARS);
         }
-        if (isset($_POST['content']) && $_POST['content'] != "") {
-            $filteredContent = filter_input(INPUT_POST, 'content', FILTER_SANITIZE_SPECIAL_CHARS);
+        if (isset($_POST['content'])) {
+            $filteredContent = $_POST['content'];
         }
-        if (isset($_POST['image']) && $_POST['image'] != "") {
-            $filteredImage = filter_input(INPUT_POST, 'image', FILTER_SANITIZE_SPECIAL_CHARS);
-        }
+//        filter_input(INPUT_POST, 'content', FILTER_SANITIZE_SPECIAL_CHARS);
 
         $title = $filteredTitle;
         $content = $filteredContent;
-        $image = $filteredImage;
+        $image = Post::updateFile($title);
         $req->execute();
+        if (($_POST['title'] = "") && ($_POST['content'] = "")) {
+            return "null";
+        }
 
 //upload Post image if it exists
-        if (!empty($_FILES[self::InputKey]['name'])) {
-            Post::uploadFile($name);
-        }
+//        if (!empty($_FILES[self::InputKey]['name'])) {
+//            Post::uploadFile($name);
+//        }
     }
 
     public static function add() {
@@ -86,15 +87,14 @@ class Post {
             $filteredTitle = filter_input(INPUT_POST, 'title', FILTER_SANITIZE_SPECIAL_CHARS);
         }
         if (isset($_POST['content']) && $_POST['content'] != "") {
-        $filteredContent  = $_POST['content']; }
+            $filteredContent = $_POST['content'];
+        }
 //        filter_input(INPUT_POST, 'content', FILTER_SANITIZE_SPECIAL_CHARS);
-        
-        
-
 
         $title = $filteredTitle;
         $content = $filteredContent;
-        $image = Post::uploadFile($title);;
+        $image = Post::uploadFile($title);
+        ;
         $req->execute();
 
 //upload post image
@@ -106,30 +106,60 @@ class Post {
 //die() function calls replaced with trigger_error() calls
 //replace with structured exception handling
     public static function uploadFile(string $imageFileName) {
-
+if (empty($_FILES[self::InputKey])) {
+            //die("File Missing!");
+            trigger_error("Please upload an image for this post");
+        }
         if (empty($_FILES[self::InputKey])) {
             //die("File Missing!");
             trigger_error("Please upload an image for this post");
         }
-
         if ($_FILES[self::InputKey]['error'] > 0) {
             trigger_error("Handle the error! " . $_FILES[InputKey]['error']);
         }
-
-
         if (!in_array($_FILES[self::InputKey]['type'], self::AllowedTypes)) {
             trigger_error("This File Type Not Allowed: " . $_FILES[self::InputKey]['type']);
         }
-
         $tempFile = $_FILES[self::InputKey]['tmp_name'];
         $path = "/Applications/XAMPP/xamppfiles/htdocs/WorldFoodBlog/uploads/";
         $destinationFile = $path . $imageFileName . '.jpeg';
         $imagePath = "uploads/" . $imageFileName . '.jpeg';
-
         if (!move_uploaded_file($tempFile, $destinationFile)) {
             trigger_error("Handle Error");
         }
+        //Clean up the temp file
+        if (file_exists($tempFile)) {
+            unlink($tempFile);
+        }
+        return $imagePath;
+    }
+    const AllowedTypess = ['image/jpeg', 'image/jpg',''];
+    const InputKeys = 'image';
 
+    public static function updateFile(string $imageFileName) {
+        if ($_FILES[self::InputKeys] == "") {
+            return "null";
+        }
+
+        if (empty($_FILES[self::InputKeys])) {
+            //die("File Missing!");
+            trigger_error("Please upload an image for this post");
+        }
+//        if ($_FILES[self::InputKeys]['error'] > 0) {
+//            trigger_error("Handle the error! " . $_FILES[InputKey]['error']);
+//        }
+        if (!in_array($_FILES[self::InputKeys]['type'], self::AllowedTypess)) {
+            trigger_error("This File Type Not Allowed: " . $_FILES[self::InputKeys]['type']);
+        }
+        $tempFile = $_FILES[self::InputKeys]['tmp_name'];
+        $path = "/Applications/XAMPP/xamppfiles/htdocs/WorldFoodBlog/uploads/";
+        $destinationFile = $path . $imageFileName . '.jpeg';
+        $imagePath = "uploads/" . $imageFileName . '.jpeg';
+        if (!move_uploaded_file($tempFile, $destinationFile)) {
+//            trigger_error("Handle Error");
+            return $imagePath;
+            
+        }
         //Clean up the temp file
         if (file_exists($tempFile)) {
             unlink($tempFile);
